+++ dayos-app/src/App.tsx
@@ -3,7 +3,6 @@ import { AppBootstrap } from './components/AppBootstrap'
 import { AppShell } from './components/AppShell'
 import { ResearchPage } from './pages/ResearchPage'
 import { SchedulePage } from './pages/SchedulePage'
-import { StatsPage } from './pages/StatsPage'
 import { TodayPage } from './pages/TodayPage'
 import { YouPage } from './pages/YouPage'
 
@@ -15,7 +14,6 @@ const router = createBrowserRouter([
       { index: true, element: <TodayPage /> },
       { path: 'schedule', element: <SchedulePage /> },
       { path: 'research', element: <ResearchPage /> },
-      { path: 'stats', element: <StatsPage /> },
       { path: 'you', element: <YouPage /> },
       { path: '*', element: <Navigate to="/" replace /> },
     ],